requirements.txt
-i http://mirrors.aliyun.com/pypi/simple/ --trusted-host mirrors.aliyun.com
python-gitlab
Flask
Flask_OAuthlib
Flask_SQLAlchemy
redis
smart_getenv
SQLAlchemy
webargs
Werkzeug
ipython
python-etcd
flask_mako
protobuf
grpcio
raven
PyYAML
webargs
more_itertools
tailer
